smolagents-demo/requirements.txt
smolagents[litellm]
smolagents[toolkit]
python-dotenv
